UI fixes

+++ app/src/main/java/dev/gaborbiro/investments/model/FTAssetUIModel.kt
@@ -3,6 +3,6 @@ package dev.gaborbiro.investments.model
 data class FTAssetUIModel(
     val symbol: String,
     val name: String,
-    val value: Float,
+    val value: Double,
     val currency: String,
 )
